move fitness to genetics subpackage
