+++ asapdiscovery-data/asapdiscovery/data/backend/plip.py
@@ -13,7 +13,7 @@ from asapdiscovery.data.backend.openeye import (
     oechem,
     save_openeye_pdb,
 )
-from asapdiscovery.data.fitness import parse_fitness_json, target_has_fitness_data
+from asapdiscovery.genetics.fitness import parse_fitness_json, target_has_fitness_data
 from asapdiscovery.data.metadata.resources import FINTSCORE_PARAMETERS
 from asapdiscovery.data.services.postera.manifold_data_validation import TargetTags
 from asapdiscovery.dataviz._gif_blocks import GIFBlockData
