+++ asapdiscovery-dataviz/asapdiscovery/dataviz/viz_v2/html_viz.py
@@ -2,7 +2,7 @@ import logging
 from enum import Enum
 from pathlib import Path
 
-from asapdiscovery.data.fitness import target_has_fitness_data
+from asapdiscovery.genetics.fitness import target_has_fitness_data
 from asapdiscovery.data.services.postera.manifold_data_validation import TargetTags
 from asapdiscovery.data.util.dask_utils import backend_wrapper, dask_vmap
 from asapdiscovery.dataviz.html_viz import HTMLVisualizer
