+++ asapdiscovery-docking/asapdiscovery/docking/scorer.py
@@ -8,7 +8,7 @@ import numpy as np
 import pandas as pd
 from asapdiscovery.data.backend.openeye import oedocking
 from asapdiscovery.data.backend.plip import compute_fint_score
-from asapdiscovery.data.fitness import target_has_fitness_data
+from asapdiscovery.genetics.fitness import target_has_fitness_data
 from asapdiscovery.data.schema.ligand import LigandIdentifiers
 from asapdiscovery.data.schema.target import TargetIdentifiers
 from asapdiscovery.data.services.postera.manifold_data_validation import TargetTags
