+++ asapdiscovery-docking/asapdiscovery/docking/workflows/small_scale_docking.py
@@ -2,7 +2,7 @@ from pathlib import Path
 from shutil import rmtree
 from typing import Optional
 
-from asapdiscovery.data.fitness import target_has_fitness_data
+from asapdiscovery.genetics.fitness import target_has_fitness_data
 from asapdiscovery.data.metadata.resources import master_structures
 from asapdiscovery.data.operators.deduplicator import LigandDeDuplicator
 from asapdiscovery.data.operators.selectors.mcs_selector import MCSSelector
